Update basic-phonebook.py
made it so when user types "list" all entries are listed in a more readable format. I had to get some help with gpt for this, but I do understand the code. It basically will take the difference between the name and longest name, and do some math to print dots so it looks like an old phone book. This also looks much better then just the literal dictionary being printed. I also removed the notes at the top because I don't really need those anymore.